+++ cmd/list.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/OpenCHAMI/magellan/internal/cache/sqlite"
+	"github.com/davidallendj/magellan/internal/cache/sqlite"
 	"github.com/rs/zerolog/log"
 
 	"github.com/spf13/cobra"