+++ cmd/root.go
@@ -20,7 +20,7 @@ import (
 	"os"
 	"os/user"
 
-	magellan "github.com/OpenCHAMI/magellan/internal"
+	magellan "github.com/davidallendj/magellan/internal"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"