+++ internal/collect.go
@@ -10,13 +10,13 @@ import (
 	"net/http"
 	"os"
 	"path"
-	"strings"
 	"path/filepath"
+	"strings"
 	"sync"
 	"time"
 
-	"github.com/OpenCHAMI/magellan/pkg/client"
-	"github.com/OpenCHAMI/magellan/pkg/crawler"
+	"github.com/davidallendj/magellan/pkg/client"
+	"github.com/davidallendj/magellan/pkg/crawler"
 
 	"github.com/rs/zerolog/log"