+++ internal/scan.go
@@ -10,8 +10,8 @@ import (
 	"sync"
 	"time"
 
-	urlx "github.com/OpenCHAMI/magellan/internal/url"
-	"github.com/OpenCHAMI/magellan/pkg/client"
+	urlx "github.com/davidallendj/magellan/internal/url"
+	"github.com/davidallendj/magellan/pkg/client"
 	"github.com/rs/zerolog/log"
 )